requirements.txt
# Core Django
Django==4.2.27
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3

# Configuration
django-environ==0.12.0

# Database
dj-database-url==3.1.0
psycopg2-binary==2.9.11

# Static files & Production
whitenoise==6.11.0
gunicorn==23.0.0

# Image handling (for ImageField)
pillow==12.1.0
